Create commands package and core/nap module
