+++ pyproject.toml
@@ -57,7 +57,7 @@ dev = [
 version = {attr = "xpk.core.core.__version__"}
 
 [tool.setuptools]
-packages = ["xpk", "xpk.parser", "xpk.core"]
+packages = ["xpk", "xpk.parser", "xpk.core", "xpk.commands"]
 package-dir = {"" = "src"}
 
 [tool.pyink]
